Rename a few things
For clarity. We still have quite a few single letter variable names,
some of them in Unicode, some subscipts roaming around, etc. etc., but
this is still an improvement and it cleans up the exported namespace so
that the standard Julia convention is followed (for the most part).

Variables beginning with capital ASCII characters containing an
underscore in their names, or variables in all caps are understood to
not be types.

Rather than rename the three main functions which would be kind of
annoying I've decided to simply create aliases for them along with a few
other functions. Internally, the single letters are convenient and
follow the mathematical convention, and they clear up a ton of space for
long function signatures.