Rename the first field in the wavefront types to `recap`
Yes, more renaming. This one was particularly confusing, not only did
they not match, but using just `i` is terrible and `a` should refer to
only the coefficients. `WavefrontError` has an `a` field already, but
the one in `WavefrontOutput` referred to the summary. I do remember
justifying this before, it's not like I hadn't thought of it, but in
retrospect this is the obvious move.