Offload the output to `show` and `Output`
Okay so this makes a lot more sense. Given that we have output types
it's a good idea to use them and have them ultimately handle the figure
display and output printing. This way output can be suppressed if
desired as well. This does require more information storage within the
`Output` struct, but this is not a problem as it could possibly be
useful accessing the Unicode and Zernike index fields without having to
bring up the figure.