Add a new `metrics` method for `WavefrontError`s
So first, `metrics` has generally been improved in two ways:

* `extrema` is now used to compute the maximum and minimum in one pass;
* extra computation with the piston term is now avoided.

The latter could possibly be more accurate in some edge cases due to
floating point error; in any case there's really no need to square, add,
square, subtract when we can just take a `view`. Indexing has also been
generally improved as it's now generic: even though we return and
usually expect a standard `Vector` from our fitting functions, if the
input is an `OffsetArray` of some sort all bets are off and since the
argument types are rather lax (accepting `AbstractArray` types) it's
better to just simply guard any indexing error, especially if one of the
methods is going to be exposed via the API.

Anyway, the new method is useful as its computation is not carried out
if the error function is returned instead.