Use the canonical radial coefficient formula for orders < 21
The previous floating factorial implementation was roughly twice as fast
and allocated much less compared to the recursive algorithm and was
perfectly accurate for n = 0:22. Given that integer factorials are
grabbed from a lookup table in Julia and 64-bit integers overflow around
the same degree where the floating factorial computation broke down in
accuracy then it makes sense to go with this new approach as it provides
even better speed; this now yields sub-microsecond return times for
generating the functions.

The Int64 check is there in order to avoid three different branching
implementations: factorials overflow even sooner for Int32 so for that
the function will fallback to the recursive algorithm as before; that's
fast enough anyway (on the order of microseconds) and when the
performance impact becomes significant that algorithm will be in use
anyway.