Specialize `show` for when the item is an element of a container
From what I can gather, on `display` array printing first calls 3-arg
`show` using `sprint` and falls back to 2-arg `show` for the `print`
stream if the string contains line breaks. Usually I suppose you'd
overload the 3-arg mime type `show` as the 2-arg `show` is meant to be
used for more concise output. Now, because we `display` the figure as
well we want to avoid this for collections of instances so now we check
the `:typeinfo` `IOContext` attribute and use it to bifurcate in a
stable manner.

The 2-arg `show` method is supplied as an additional provision for
collections such as `Set([w1, w2])` which seems to call 2-arg `show`
first and abbreviates / truncates if there are line breaks instead. In
this case the `:typeinfo` is missing so we now provide a custom concise
print stream as well.

Also, construction of the `inds` string has been simplified as it's
somewhat related since it's also used for `Output` printing and it
doesn't really deserve its own commit.