Define default value constants for the keyword argument options
I'm not sure why I didn't do this before given how standard and basic a
practice this is and considering there are other less widely used
constants already in there; it would've been less work and much safer
had I defined these at the very beginning so I'm not sure what I was
thinking. Obviously this is much easier to manage now since only one
value needs to be changed in order to change the default setting across
the board.