Define separate methods for the three public coefficient functions
This specifies things more clearly, although now the names are kinda
long, but that's not really a big deal as these functions aren't meant
to be used frequently and this avoids confusion; if they are to be used
they will probably be used in an automated fashion.

What this also does is instead of just creating aliases these distinct
functions now directly access the desired quantities.