Define some methods for working with reduced `precision`
Since `precision` is an `Int` the constructor needs to be called,
otherwise a mutating `reduce_wave!` could be defined and the three
vectors mutated directly. Maybe some day I'll make `WavefrontError` a
mutable struct instead.

`sieve` is now there for convenience, and the new `standardize` method
makes it straightforward to return the full vector corresponding to the
reduced `a`.