Define `WavefrontError(Z::Polynomial)`
This is the preferred method for converting Zernike polynomials.

I'm not sure whether equality is best defined right now by comparing
every field as `WavefrontError(Z) == 1.0 * Z` returns false since the
LHS has precision 3 and the RHS has precision 17 by default and since
the fields (and not the properties) are being compared then fit_to also
makes a difference. So the coefficient vectors could be identical, but
the wavefront comparison will currently return false.

In a way this is desirable, as computations using these wavefronts could
differ and so they're practically different, but at the same time
they're essentially identical. Since it's not possible to define the
built-in egal function (===) for custom types (since it's probably not a
good idea to do so) then this will have to do; it's simple enough to
execute W1.a == W2.a anyway.