TransformAperture: accept non-standard length coefficient vectors
So I'm not going to change the documentation here because it's probably
a good idea to remain explicitly clear about what the algorithm requires
and it could be sort of confusing having the transformation function
return a vector of a different length than what was input if you don't
expect it.

That being said, this is convenient and avoids having to pad the vector
yourself.