Use simpler names for the Cartesian pupil coordinates
I've been using rho_{x,y} following convention where the x and y symbols
are reserved for the spatial coordinates of an optical system under an
extended description and so the pupil coordinates need to be
distinguished. We don't have that here, so there's no need to keep
following that when it makes things more confusing, more difficult to
read, and the non-LaTeX variable names were being subscripted with chi
and gamma instead anyway; although I suppose having unique names does
make things easier to find in a quick search, but that's entirely not
necessary for these.