Extend the set of suitable coefficient indexing methods
- Vector indices are now supported for getindex;
- lastindex allows for the convenient z[end];
- firstindex is there for symmetry and to signal the zero-based system;
- setindex! for WavefrontErrors is a convenient method for directly
  changing the coefficients using powers for the indices.

AbstractPolynomial types are standardized and should not support
directly changing the coefficients.

Note that in order for broadcasting to work `Base.dotview` from the
Broadcast module needs to be overloaded as that is what `w[vec] .= a`
lowers to, but at the moment that is not a public function and internal
methods like that are subject to change in Julia; also it has no
documentation.

No further unit tests are planned for these implementations as they are
really meant to be syntactic sugar for use on reflection.