Build documentation using Documenter
The README is getting a bit long so some proper docs should make things
more navigable. That being said it's not *that* long and I kinda like
everything in one place & on the front page so since the README is
pretty comprehensive the docs are simply being generated from it by
splitting and post-processing it; having this automated at build time
obviates the need to maintain both the README and the docs. If the
documentation keeps getting longer though I might just shorten the
README down to a précis as is customary.

The build script is supposed to be robust enough to handle a variety of
document restructuring and also support local builds which is why the
script is somewhat longish and the logic somewhat involved.

Also, the API reference is quite useful which would have been the
primary reason to build docs for this package in the first place.

Building the docs caught a few dormant bugs mostly due to typos so those
have been fixed as well:

- the removed standardize! method was still being referenced;
- coefficients --> radial_coefficients in Docstrings;
- plotconfig --> Zernike.plotconfig in the manual;
- wavefront_coefficents --> wavefront_coefficients.

`Standard` & `wavefront_coefficients` are now also documented as well.