Expose the `plotconfig` reset methods and remove the `setproperty!` hook
Although it was kind of cool having the functionality live through
setting pseudo-properties it was also a bit more awkward, non-standard,
and also more code. getproperty also didn't work on those two.

The methods have to be qualified with the module name so maybe that's a
bit more annoying to type compared to `= true` and if the plotconfig
isn't imported maybe this isn't an improvement, but I figure if the
settings are going to be changed frequently the config will probably be
imported. resize! & reset! are common names though, the first has a Base
method and the other has one in Makie and it would be kind of awkward to
overload just one.

Either way this is better style and it's easier to remember especially
given that resetting settings isn't likely to be needed frequently.