Support extraction of the Unicode string representation for derivatives
I tried to keep the changes to format_strings minimal so as not to break
anything, especially since it's not the easiest code to follow, but
there might be a bug in there; I did manually test it of course.

The string formatting test suite is quite sparse at the moment, so I
won't be adding any for these, rather maybe I'll add a bunch at some
later date, but right now it's not a priority.