Simplify the computation of derivatives
So I realized you could construct a differentiation operator by
specifying the superdiagonal of a sparse matrix so this should also be
highly efficient.

For the sign of the harmonic derivatives what I came up with before was
not computationally expensive, but it was difficult to understand at a
glance; this is simplified by recognizing the fact that the sign goes as
(-1) ^ k where k is a triangular number depending on the order of the
derivative and the sign of m.