Rename the pupil transform function `P` to `Y`
This is technically breaking, but I plan to release a major version soon
and I don't think anyone but me uses this package anyway. Unlike the Z
and W functions this one was always arbitrary since it doesn't
have a standard analog in the literature.

I'm already using P to denote products in the tests, and it could also
be confused with phases or just polynomials. The main reason I'm
changing this now is because I plan to add some features that use Jacobi
polynomials and as those use P I'd like to follow the standard notation.