Docs: add a few reference links to the API, include meta blocks
GitHub doesn't have an issue rendering the jldoctest code block in the
README, I suppose it doesn't recognize it as a valid language and just
doesn't apply the appropriate syntax highlighting; that's fine because
it's meant for Documenter.

As for the plot example, if that line isn't suppressed the following
line needs to be added to the meta block as an inverse check for the
doctest not to fail:

DocTestFilters = r"^[^√]*\$"

But that's too specific and too much of a workaround. As it is it's fine
explaining that plotting is in fact not on demand as `display` needs to
be called.