Add a Laplacian constructor and a Cartesian `Gradient` method
The Laplacian computes an extra derivative and consumes an extra
function call, but this does seem like the most straight-forward way to
do it and these computations should be fast anyway; I can always
optimize and simplify it later. This is wholly untested for now, there
are a few more features I want to add before validating this.