Declare some fields holding parametric types as concrete
It's not a great idea to have parametric abstract types floating around
like that when not necessary; in any case being explicit here makes
things clear.

WavefrontError has a type parameter not out of necessity, but because
it is structurally pleasant and because this design makes things easily
extensible; it's easy to imagine a WavefrontError holding derivative
types, for example, and extending the arithmetic for that. I am fixing
things here, however, as I won't be adding that given there's a relation
between derivatives and the standard Zernike basis, so any wavefront
conversion will simply use that.

The value of the Vector{T} field was fine since Polynomial has no
subtypes and is concrete, but like I said this makes it easier to read
and understand at a glance since the element type of that vector is in
fact fixed.

For WavefrontOutput I hadn't specified a type parameter so the field was
declared abstract as WavefrontError is a UnionAll so maybe I wasn't
making it easy on the compiler, I don't know.

As for MixedPhase there may have been a type instability I missed
somewhere and for computationally heavy items like this it's better to
anchor things.