grad: fix a comment, add a helper method for complex vector extraction
This should be made explicitly clear in that the column order of the
matrix will not vary depending on the sign of the azimuthal order;
the conjugate derivatives are computed in a fixed fashion so care must
be taken to extract the appropriate vectors from the matrix depending on
which standard real-valued polynomial you're actually differentiating.

The helper method is there to do that -- it's cleaner to just return the
entire complex matrix at main grad and call this one when it's needed
(e.g. requiring the specific complex coefficients outright). I'm not
sure I'll need this for when I implement wavefront derivative methods
(i.e. computing the wavefront error coefficients given wavefront error
derivatives -- also from the same Janssen paper), but this should be
there for completeness.