ZernikePlot: use the low level `LaTeXString` constructor for log titles
Without this the entire string is wrapped in LaTeX math mode anchors `$`
if plot_title is normal text so the entire thing is treated as if it's a
formula or expression and thus the spaces are ignored.

Constructing the LaTeXString by concatenating the plain string with the
plot_title will ensure that `$`s are not inserted and thus the entire
title will render correctly regardless if plot_title is a normal string
or a LaTeXString.