Clarify `wavefront_finesse`
I was going to refactor this, but I remembered why I designed it this
way so I'm adding a comment to make things clear.

Usually you'd set the default in the signature using a function or so,
e.g. kwarg = f(x), but here we have multiple functions splitting kwargs
and then calling the synth function so the kwargs can't be directly
forwarded. There are other ways to approach this, but I find it easier
to just do it this way and handle it at synth; this way if a non-valid
finesse is entered there's also a fallback value.