Properly implement `setindex!` for `Wavefront`
These two vectors also need to be updated, especially `a` since those
are the coefficients which are used to compute values over the pupil.

Maybe this shouldn't be allowed, given how the fields are somewhat
dependent; once you update a value the wavefront is no longer the result
of fitting, but of course Wavefronts are not necessarily always the
result of fitting so given the general manipulations allowed this fits
the theme and this scheme can be quite useful.