ZernikePlot: pass kwargs onto `surface!`, define a shading constant
Any additional keyword arguments should be forwarded to the plotting
function to allow for more fine grained control. Shading is off by
default, due to how the plots serve a dual role as surface plots and
essentially heatmap-like images using a top-down view, and also because
the shape of the matrices is length(θ)-by-length(ρ) with the usual axes
reversed so the lighting comes from below. This can now be overridden
using kwargs with: shading = true, invert_normals = true.