ZernikePlot: set the default theme to black
To me this looks better. The default white background is too bright on
most modern displays, and theme_dark() doesn't have enough contrast.

activate! and set_theme! are being called within zplot rather than
outside of it & on config changes because the user can change these
settings between plots, when plotting other things, for
example, and so we need to make sure the default Zernike plot options
are respected; also setting up listeners just for the theme &
focus_on_show would not be a great solution.

Now that I think about it, I might just change this to using
`with_theme` instead so that the theme resets since forcing a theme is
also not a great solution given that it also affects any subsequent
plots that are not necessarily Zernike plots.