ZernikePlot: only set the default theme for package specific plots
So, finally doing this properly. Now I kind of have to given that the
package is setting visual themes.

The screen_config options are global, so there seems to be no way around
that aspect, which is why I've included a switch to reset it within our
reset! function. The theme is the most intrusive in a way anyway, and
the focus_on_show can be set through plotconfig or kwargs and if a plot
is created then that'll reset that.

I can't decide which is more annoying: an interactive plot window being
created unfocused, or it taking focus from the command line. I figure
it's probably easier to switch back to the REPL than to switch to the
plot on most systems. Anyway, the option is there to disable it.
I personally prefer it this way (for this package) which is why the
default is what it is, and given I'm likely the only person actively
using this then I'll leave the option with the current default.

Also, plotconfig is now exported.