ZernikePlot: improve the type stability of the internal _zplot function
So apparently some of these weren't being inferred by the compiler, and
since we're calling getindex and getproperty on them I think it's better
to use typed declarations for them.

I doubt there is a noticeable improvement in speed, and I didn't bother
benchmarking this time, but of course I tested it and the plotting still
works. The types of the values bound to these variables are practically
guaranteed to not change, and there shouldn't be any conversion either,
so they also effectively act as kinds of assertions since an exception
will be thrown if conversion isn't possible.

Since we store the surface plot type in one of the output structs that
type is constrained anyway, so the assertion on that very specific
parametric type won't break anything that wouldn't break anyway if
conversion fails in case Makie's surface type changes again.

Also, the z-axis attributes can be declared constant outside of the
function obviously.