Add a tool which constructs a matrix of m, n indices and an input vector
This is not meant to be efficient given the abstract type parameter,
it's mostly for inspection / reflection and working with the rows or
columns when needed.

Now that I'm writing this commit I realize there's a simpler way to
write the function, instead of amending I'll just commit it next.