Update documentation with the new derivative algorithms
The API here is subject to change, but to me it's intuitive. Having
equivalent methods might be confusing though, but for now I like it this
way.

I've also added a `lap` function for symmetry with `grad`.

Part of the reason the multiple methods are there even though having
Gradient and grad, Laplacian and lap can be confusing, is because having
Wavefront constructor methods for them makes sense especially since they
return a Wavefront, but also because it's nice having shorthand
effective aliases; it should be easy to remember which method to use
since Gradient and Laplacian return instances of their type then by
deduction grad and lap should refer to the Wavefront constructors.

What might be confusing is having grad(m, n) and grad(Z) do different
things, as it's tempting to call grad(m, n) rather than grad(Z(m, n)),
but again this provides symmetry with Gradient(Z) and there are the
Wavefront methods in case this is confusing.

Also, the kwargs for some of these methods have been removed as they
expect normalized polynomials and commit 82236e2 has been partially
reverted; maybe I should export the derivative names, but these are
common.