Attempt to fix the MMT tests
This still fails intermittingly, even when comparing the vector norms
and using a low absolute tolerance. The algorithm seems to work for the
most part, but there's some weird numerical instability bug which causes
it to blow up at certain points which shouldn't happen because the
random coefficients and the radial evaluation points are both between
zero and one. I suspect this might have to do with how the wavefront is
constructed. I don't know, I'll come back to this when I finish the rest
of the commits.