Add a few MMT tests for negative values of m
As for the tests previously failing on large(r) numbers of samples, I
think it has to do with the product expansion yielding polynomials with
very large radial coefficients resulting in some floating point
cancellation issue; so even though the expansion coefficients could be
very small (effectively zero) and we're evaluating on the unit disk
there's a great loss of accuracy.

This should be fixable to some extent with some appropriate filtering by
reintroducing sieving of the coefficients, but the problem with that is
that the loss of precision there will also lead to a loss of accuracy,
so there seems to be a delicate balance; I'll deal with this in the next
commit, but the number of samples and tolerances will be pretty
arbitrary without some rigorous numerical analysis.