Small changes with regards to broadcasting in `standardize` & `Noll`
For standardize(fringe) elementwise assignment should be better since
it's computing something on the right hand side so this avoids
materializing an intermediate array.

For standardize(v_sub) using setindex! is clearer syntax since v_sub is
guaranteed to be an array and not a scalar making the intention of
replacing a selection with a vector more obvious; I don't think there's
any performance difference here since the identity broadcast overhead on
the rhs should be completely negligible and the compiler should optimize
it away. This one's simply a conventional change.

For the Noll constructor, there's no need to allocate an extra j array
since it's not used anywhere except in j_to_noll so we can take
advantage of fusion here.