Change a radial coefficient test
Since commit fbe2fd1 changed the expected input this needs to be changed
so that it doesn't fail. There's no need to require non-negative values
of m given that we can just take the absolute value. There's no place
this function is called with negative values, but without taking the
absolute value it should be validated; although redundant calling abs
twice (once at the call site) costs practically nothing and either way
this is done outside of any loops (at the call site and in the radial
function). This is simply better design and it allows the function to be
called with direct indices for any azimuthal order.