Fix type instabilities on overloaded `getproperty`
I found that wrapping getproperty calls in a function such as f(z) = z.m
and then using the code_warntype macro or the inferred macro from Test
showed that two of our types were unstable on property access, which is
not so great if this is done in hot loops.

As far as I understand, the recursion due to overloaded getproperty
calling getproperty again led to the same path being entered and for
some reason the condition inside our method couldn't be proved at
compile time meaning the branch wasn't being pruned preventing constant
propagation of the Symbol thus leading to the return type being widened
and so dynamic dispatch and a runtime look-up. Using getfield directly
allows the compiler to propagate the field name literal and inline the
function leading to inference succeeding. In any case, the return type
does seem to get determined now.

Note, the overload at Arithmetic for non-Radial Wavefronts seems fine.
This overload is only in effect under the hacky conversion scheme for
Polynomial --> Wavefront{Polynomial}, the approach of which isn't so
great, but this method is type stable on access so I'm going to leave it
for now until I figure out a better way to carry through the precision.