Docstrings: qualify references to unexported names in redirected string
Apparently the double doc alias redirection macro line at the end of the
file was causing documentation builds to fail sporadically.

I cherry-picked the last commit onto the base branch and it succeeded,
but locally and on dev Documenter was trying to resolve Main.name for
these names. There's something strange with how the macro is evaluated
in the module and somehow the references change scope when Documenter
looks at it and can't find the docstring. There's definitely something I
don't understand about this, it's an utterly bizarre bug, but qualifying
these three seems to fix the issue.