Wavefront: generalize the main default constructor
Julia defines default non-parametric constructors for parametric types
with restricted arguments, so this method can't be deleted without
breaking things as its called with less specific types in several
places and that's been supported thus far in order to allow for more
general construction; that's not a problem as conversion happens at
instantiation.

However, instead of forcing a non-parameterized call to route through to
the Polynomial method we can generalize this to allow for
RadialPolynomial construction without having to apply the type.
Specializations for RadialPolynomial remain explicit as their
implementations are specific to them.